+++ ext/PigeonsBridgeStanExt/PigeonsBridgeStanExt.jl
@@ -1,0 +1,28 @@
+module PigeonsBridgeStanExt
+
+using Pigeons
+if isdefined(Base, :get_extension)
+    using BridgeStan
+    using LogDensityProblems
+    using Serialization
+    using LogDensityProblemsAD
+    using DocStringExtensions
+    using Random
+else
+    using ..BridgeStan
+    using ..LogDensityProblems
+    using ..Serialization
+    using ..LogDensityProblemsAD
+    using ..DocStringExtensions
+    using ..Random
+end
+
+import Pigeons: StanLogPotential
+
+
+
+include(joinpath(@__DIR__, "interface.jl"))
+include(joinpath(@__DIR__, "state.jl"))
+include(joinpath(@__DIR__, "toy_stan_target.jl"))
+
+end
