+++ src/explorers/SliceSampler.jl
@@ -7,16 +7,16 @@ $FIELDS
 """
 @kwdef struct SliceSampler
     """ Initial slice size. """
-    w::Float64 = 10.0 
+    w::Float64 = 10.0
 
     """ Slices are no larger than 2^p * w """
-    p::Int = 20 
+    p::Int = 20
 
     """ Number of passes through all variables per exploration step. """
-    n_passes::Int = 3  
+    n_passes::Int = 3
 
     """ Maximum number of interations inside shrink_slice! before erroring out """
-    max_iter::Int = 4_096  
+    max_iter::Int = 4_096
 end
 
 explorer_recorder_builders(::SliceSampler) = [explorer_acceptance_pr, explorer_n_steps]
@@ -30,9 +30,9 @@ function step!(explorer::SliceSampler, replica, shared)
 end
 
 function cached_log_potential(log_potential, state, cached_lp)
-    return if cached_lp == -Inf 
+    return if cached_lp == -Inf
         result = log_potential(state)
-        if result == -Inf 
+        if result == -Inf
             error("SliceSampler supports contrained target, but the sampler should be initialized in the support: $state")
         end
         return result
@@ -44,32 +44,14 @@ end
 function slice_sample!(h::SliceSampler, state::AbstractVector, log_potential, cached_lp, replica)
     cached_lp = cached_log_potential(log_potential, state, cached_lp)
     # iterate over coordinates
-    for c in 1:length(state) 
+    for c in 1:length(state)
         pointer = Ref(state, c)
         cached_lp = slice_sample_coord!(h, replica, pointer, log_potential, cached_lp)
     end
     return cached_lp
 end
 
-function slice_sample!(h::SliceSampler, state::DynamicPPL.TypedVarInfo, log_potential, cached_lp, replica)
-    cached_lp = cached_log_potential(log_potential, state, cached_lp)
-    for i in 1:length(state.metadata)
-        for c in 1:length(state.metadata[i].vals)
-            pointer = Ref(state.metadata[i].vals, c)
-            cached_lp = slice_sample_coord!(h, replica, pointer, log_potential, cached_lp)
-        end
-    end
-    return cached_lp
-end
 
-function slice_sample!(h::SliceSampler, state::StanState, log_potential, cached_lp, replica)
-    cached_lp = cached_log_potential(log_potential, state, cached_lp)
-    for i in eachindex(state.unconstrained_parameters)
-        pointer = Ref(state.unconstrained_parameters, i)
-        cached_lp = slice_sample_coord!(h, replica, pointer, log_potential, cached_lp)
-    end
-    return cached_lp
-end
 
 function slice_sample_coord!(h, replica, pointer, log_potential, cached_lp)
     rng = replica.rng
@@ -84,12 +66,12 @@ function slice_sample_coord!(h, replica, pointer, log_potential, cached_lp)
 end
 
 function Bernoulli_sample_coord!(replica, pointer, log_potential, cached_lp)
-    state = replica.state 
+    state = replica.state
     rng = replica.rng
     if pointer[] == Bool(0)
         lp0 = cached_lp
         pointer[] = Bool(1)
-        lp1 = log_potential(state) 
+        lp1 = log_potential(state)
     else
         lp1 = cached_lp
         pointer[] = Bool(0)
@@ -111,14 +93,14 @@ function slice_double(h::SliceSampler, replica, z, pointer, log_potential)
     old_position = pointer[] # store old position while avoiding memory allocation
     L, R = initialize_slice_endpoints(h.w, pointer, rng, typeof(pointer[])) # dispatch on either float or int
     K = h.p
-    
+
     pointer[] = L
     potent_L = log_potential(state) # store the log potential
     pointer[] = R
     potent_R = log_potential(state)
 
     while (K > 0) && ((z < potent_L) || (z < potent_R))
-        V = rand(rng)        
+        V = rand(rng)
         if V <= 0.5
             L = L - (R - L)
             pointer[] = L
@@ -130,7 +112,7 @@ function slice_double(h::SliceSampler, replica, z, pointer, log_potential)
         end
         K = K - 1
     end
-    @record_if_requested!(replica.recorders, :explorer_n_steps, (replica.chain, h.p - K)) 
+    @record_if_requested!(replica.recorders, :explorer_n_steps, (replica.chain, h.p - K))
 
     pointer[] = old_position # return the state back to where it was before
     return (L, R, potent_L, potent_R)
@@ -145,7 +127,7 @@ end
 function initialize_slice_endpoints(width, pointer, rng, ::Type{T}) where T <: Integer
     width = convert(T, ceil(width))
     L = pointer[] - rand(rng, 0:width)
-    R = L + width 
+    R = L + width
     return (L, R)
 end
 
@@ -161,13 +143,13 @@ function slice_shrink!(h::SliceSampler, replica, z, L, R, lp_L, lp_R, pointer, l
 
     while n <= h.max_iter
         new_position = draw_new_position(Lbar, Rbar, rng, typeof(pointer[]))
-        pointer[] = new_position 
+        pointer[] = new_position
         new_lp = log_potential(state)
-        consider = z < new_lp 
+        consider = z < new_lp
         pointer[] = old_position
         if consider && slice_accept(h, replica, new_position, z, L, R, lp_L, lp_R, pointer, log_potential)
             pointer[] = new_position
-            @record_if_requested!(replica.recorders, :explorer_n_steps, (replica.chain, n)) 
+            @record_if_requested!(replica.recorders, :explorer_n_steps, (replica.chain, n))
             return new_lp
         end
         if new_position < pointer[]
@@ -200,14 +182,14 @@ function slice_accept(h::SliceSampler, replica, new_position, z, L, R, lp_L, lp_
     # tracks whether lp_R,lp_L need to be recomputed
     Rstale = false
     Lstale = false
-    
+
     D = false
     while Rhat - Lhat > 1.1 * h.w
         M = (Lhat + Rhat)/2.0
         if ((old_position < M) && (new_position >= M)) || ((old_position >= M) && (new_position < M))
             D = true
         end
-        
+
         if new_position < M
             Rhat = M
             Rstale = true
@@ -228,7 +210,7 @@ function slice_accept(h::SliceSampler, replica, new_position, z, L, R, lp_L, lp_
                 Rstale = false
             end
             if  ((z >= lp_L) && (z >= lp_R))
-                pointer[] = old_position 
+                pointer[] = old_position
                 @record_if_requested!(replica.recorders, :explorer_acceptance_pr, (replica.chain, 0.0))
                 return false
             end
