+++ src/includes.jl
@@ -74,11 +74,9 @@ include("explorers/hamiltonian_samplers.jl")
 include("explorers/Compose.jl")
 include("explorers/Augmentation.jl")
 include("targets/TuringLogPotential.jl")
-include("targets/toy_turing_target.jl")
 include("explorers/BufferedAD.jl")
 include("variational/GaussianReference.jl")
 include("targets/StanLogPotential.jl")
-include("targets/toy_stan_target.jl")
 include("pt/checks.jl")
 include("paths/ScaledPrecisionNormalPath.jl")
 include("targets/toy_mvn_target.jl")
