+++ src/targets/toy_mvn_target.jl
@@ -1,13 +1,13 @@
-""" 
-$SIGNATURES 
-
-A toy multi-variate normal (mvn) target distribution used for testing. 
-Uses a specialized path, [`ScaledPrecisionNormalPath`](@ref), 
-such that i.i.d. sampling is possible at all chains (via [`ToyExplorer`](@ref)). 
 """
-@provides target toy_mvn_target(dim::Int) = ScaledPrecisionNormalPath(dim) 
+$SIGNATURES
 
-initialization(target::ScaledPrecisionNormalPath, rng::AbstractRNG, _::Int64) = 
+A toy multi-variate normal (mvn) target distribution used for testing.
+Uses a specialized path, [`ScaledPrecisionNormalPath`](@ref),
+such that i.i.d. sampling is possible at all chains (via [`ToyExplorer`](@ref)).
+"""
+@provides target toy_mvn_target(dim::Int) = ScaledPrecisionNormalPath(dim)
+
+initialization(target::ScaledPrecisionNormalPath, rng::AbstractRNG, _::Int64) =
     randn(rng, target.dim) / sqrt(target.precision1)
 
 default_explorer(::ScaledPrecisionNormalPath) = ToyExplorer()
@@ -19,6 +19,3 @@ Random.rand!(rng::AbstractRNG, x::AbstractVector, log_potential::ScaledPrecision
     for i in eachindex(x)
         x[i] = randn(rng) / sqrt(log_potential.precision)
     end
-
-Random.rand!(rng::AbstractRNG, state::Pigeons.StanState{Vector{Float64}}, log_potential::Pigeons.ScaledPrecisionNormalLogPotential) = 
-    rand!(rng, state.unconstrained_parameters, log_potential)
