+++ test/supporting/setup.jl
@@ -13,7 +13,9 @@ using SplittableRandoms
 using Statistics
 using Test
 using ForwardDiff
-using LogDensityProblems 
+using LogDensityProblems
 using LogDensityProblemsAD
+using DynamicPPL
+using BridgeStan
 
 is_windows_in_CI() = Sys.iswindows() && (get(ENV, "CI", "false") == "true")
